+++ src/ScottPlot5/ScottPlot5/Panels/ColorBar.cs
@@ -97,7 +97,7 @@ public class ColorBar : IPanel
         axis.Min = range.Min;
         axis.Max = range.Max;
 
-        axis.TickGenerator.Regenerate(axis.Range, axis.Edge, length);
+        axis.TickGenerator.Regenerate(axis.Range.ToCoordinateRange, axis.Edge, length);
 
         return axis;
     }