+++ src/ScottPlot5/ScottPlot5/Primitives/CoordinateRangeStruct.cs
@@ -1,7 +1,0 @@
-﻿namespace ScottPlot;
-
-public readonly record struct CoordinateRangeStruct(double Min, double Max)
-{
-    public double Span => Max - Min;
-    public double Center => (Min + Max) / 2;
-}