+++ src/ScottPlot5/ScottPlot5/Rendering/RenderActions/EnsureAxesHaveArea.cs
@@ -7,7 +7,7 @@ public class EnsureAxesHaveArea : IRenderAction
 {
     public void Render(RenderPack rp)
     {
-        foreach (CoordinateRange range in rp.Plot.Axes.GetAxes().Where(x => x.Range.Span == 0).Select(x => x.Range))
+        foreach (CoordinateRangeMutable range in rp.Plot.Axes.GetAxes().Where(x => x.Range.Span == 0).Select(x => x.Range))
         {
             range.Min -= 1;
             range.Max += 1;