+++ src/ScottPlot5/ScottPlot5/Rendering/RenderActions/RegenerateTicks.cs
@@ -4,7 +4,14 @@ public class RegenerateTicks : IRenderAction
 {
     public void Render(RenderPack rp)
     {
-        rp.Plot.Axes.Bottom.TickGenerator.Regenerate(rp.Plot.Axes.Bottom.Range, rp.Plot.Axes.Bottom.Edge, rp.DataRect.Width);
-        rp.Plot.Axes.Left.TickGenerator.Regenerate(rp.Plot.Axes.Left.Range, rp.Plot.Axes.Left.Edge, rp.DataRect.Height);
+        rp.Plot.Axes.Bottom.TickGenerator.Regenerate(
+            range: rp.Plot.Axes.Bottom.Range.ToCoordinateRange,
+            edge: rp.Plot.Axes.Bottom.Edge,
+            size: rp.DataRect.Width);
+
+        rp.Plot.Axes.Left.TickGenerator.Regenerate(
+            range: rp.Plot.Axes.Left.Range.ToCoordinateRange,
+            edge: rp.Plot.Axes.Left.Edge,
+            size: rp.DataRect.Height);
     }
 }